requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

pixivpy == 3.7.0